requirements.txt
aiohttp==3.9.1
python-telegram-bot==20.7
asyncio==3.4.3
requests==2.31.0
python-dotenv==1.0.0
